asx_bot/requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
keras>=3.0.0
tensorflow>=2.15.0
python-dotenv>=1.0.0
ib-insync>=0.9.86
flask>=3.0.0
requests>=2.31.0
